изменен ти переменных

+++ WindowsFormsApplication1/Form1.cs
@@ -18,9 +18,9 @@ namespace WindowsFormsApplication1
 
         private void button1_Click(object sender, EventArgs e)
         {
-            int a = Convert.ToInt32(textBox1.Text);
-            int b = Convert.ToInt32(textBox2.Text);
-            int c;
+            float a = Convert.ToSingle(textBox1.Text);
+            float b = Convert.ToSingle(textBox2.Text);
+            float c;
             c = a + b;
             textBox3.Text = Convert.ToString(c);
             label4.Text = "Сложение";
@@ -39,9 +39,9 @@ namespace WindowsFormsApplication1
 
         private void button2_Click(object sender, EventArgs e)
         {
-            int a = Convert.ToInt32(textBox1.Text);
-            int b = Convert.ToInt32(textBox2.Text);
-            int c;
+            float a = Convert.ToSingle(textBox1.Text);
+            float b = Convert.ToSingle(textBox2.Text);
+            float c;
             c = a - b;
             textBox3.Text = Convert.ToString(c);
             label4.Text = "Вычитание";
@@ -50,9 +50,9 @@ namespace WindowsFormsApplication1
 
         private void button3_Click(object sender, EventArgs e)
         {
-            int a = Convert.ToInt32(textBox1.Text);
-            int b = Convert.ToInt32(textBox2.Text);
-            int c;
+            float a = Convert.ToSingle(textBox1.Text);
+            float b = Convert.ToSingle(textBox2.Text);
+            float c;
             c = a * b;
             textBox3.Text = Convert.ToString(c);
             label4.Text = "Умножение"; 
@@ -61,9 +61,9 @@ namespace WindowsFormsApplication1
 
         private void button4_Click(object sender, EventArgs e)
         {
-            int a = Convert.ToInt32(textBox1.Text);
-            int b = Convert.ToInt32(textBox2.Text);
-            int c;
+            float a = Convert.ToSingle(textBox1.Text);
+            float b = Convert.ToSingle(textBox2.Text);
+            float c;
             c = a / b;
             textBox3.Text = Convert.ToString(c);
             label4.Text = "Деление";
